+++ fabric/src/main/java/org/infernalstudios/questlog/QuestlogFabric.java
@@ -8,12 +8,14 @@ import org.infernalstudios.questlog.networking.QuestlogPacketsFabric;
 public class QuestlogFabric implements ModInitializer, ClientModInitializer {
     @Override
     public void onInitialize() {
-        QuestlogPacketsFabric.register();
+        QuestlogPacketsFabric.registerCommon();
         QuestlogFabricEventForwarder.init();
     }
 
     @Override
     public void onInitializeClient() {
+        QuestlogPacketsFabric.registerClient();
+
         Questlog.initClient();
         KeyBindingHelper.registerKeyBinding(QuestlogClient.OPEN_SCREEN_KEY);
         QuestlogFabricEventForwarder.initClient();